Rename neo to neonix to enable standalone import
The initial decision to structure the repository to match the structure
of python-neo causes issues when attempting to import the nixio
independently. It becomes difficult (impossible?) to import both
python-neo and our IO since both packages are named neo. This rename
allows us to treat "neonix" as a separate package. The package will be
renamed and all import will be (trivially) refactored before finalising
the IO for integration into python-neo.