Create name for nix object if neo name is None
Names are created as "neo.ObjectType#", where # is an incremental number
of nix equivalent/mapped objects. This means that for certain objects,
such as AnalogSignal and IrregularlySampledSignal, which both create
DataArrays, the incremental number is shared.
The number isn't meant as a index of each object, it is simply there to
guarantee uniqueness of names.