Object path redesign
The path variable, used in write methods for accessing parent objects
and creating metadata structures, has been redesigned.
It is now a '/' delimited string of object names and Neo containers.
The _get_object_at method along with the _container_map dictionary
handle the translation from a path string to a specific NIX object.

The redesign is based on the HFD5 IO in Neo. This change will make
loading objects straightforward for the reader as well as the API user.
The use of Neo paths hides the NIX structure from the Neo user, making
the IO a proper IO for Neo.